update image so it uses size_t instead of int for dimensions
git-svn-id: file:///home/roberto/Git/svn@6509 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 131896561e122976e2294218eb9b2f78373d5c61

+++ trunk/xmipp/libraries/data/memory.cpp
@@ -1,5 +1,5 @@
 #include "memory.h"
-char*  askMemory(unsigned long memsize)
+char*  askMemory(size_t memsize)
 {
     char*		ptr = NULL;
